TALASH_Milestone2/talash/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pdfplumber==0.11.4
groq==0.11.0
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
streamlit==1.39.0
plotly==5.24.1
httpx==0.27.2
pydantic==2.9.2
aiofiles==24.1.0
